requirements.txt
google-genai
easyocr
opencv-python
pillow
tqdm
numpy
